backend/agent_backend/pyproject.toml
[project]
name = "agent_backend"
version = "0.1.0"
description = "agent_backend using crewAI"
authors = [{ name = "Your Name", email = "you@example.com" }]
requires-python = ">=3.10,<3.14"
dependencies = [
    "crewai[tools,google-genai]==1.11.0",
    "opik>=1.10.58",
    "python-dotenv>=1.0.0,<2.0.0",
]

[project.scripts]
agent_backend = "agent_backend.main:run"
run_crew = "agent_backend.main:run"
match = "agent_backend.main:run_matching_cli"
convo = "agent_backend.main:run_conversation_cli"
train = "agent_backend.main:train"
replay = "agent_backend.main:replay"
test = "agent_backend.main:test"
run_with_trigger = "agent_backend.main:run_with_trigger"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.crewai]
type = "crew"
